Remove duplicate scripts in favor of /etc/hosts
Previously, switching to a different network environment (such as when
moving between home and internship housing) would require changes to the
following files in order to make all Odroid server services accessible:

- /etc/fstab (for Samba mounts)
- /usr/local/bin/restic-bak.sh (for backups)
- /usr/local/bin/NAS-privdata_mount.sh (for remote drive mounting)
- ~/.ssh/config (for SSH access, new environments only)
- /root/.ssh/config (for backups and remote drive mounting, new
  environments only)

I maintained a copy of each of these files for every environment (for
example, /etc/fstab-home.bak and /etc/fstab-school.bak) that I could
copy over to switch instead of manually editing the SSH host/IP address used.
However, that was still not as quick of a process as it can be. This
change removes those duplicate files in favor of a single file which
uses a hostname specified in /etc/hosts. This means there is now a
single point of change for switching between network environments, which
is just changing the IP address for the alarm.localdomain hostname in
/etc/hosts.