Set USB audio as default sink when connected
Previously, any active audio stream would switch to playing on the new
audio device (such as my new USB-C hub) when plugged in (as intended).
However, because the USB audio wasn't the default sink, the volume keys
still controlled the built-in audio. This commit ensures that the volume
keys will automatically switch to controlling the USB audio when plugged
in.

Note that by default this setting applies to all new sinks and sources
(not just the audio sink for my USB hub) except for HDMI devices.