Avoid root use of pulseaudio for microphone toggle
The most recent round of updates appears to have changed something in
the Arch Linux audio system which broke the old setup. Inspection showed
that this was because root (which is used because it has the appropriate
permissions to set the keyboard LEDs) was seeing different simple mixer
controllers listed in the output of amixer, so changes made by my user
account to the Capture controller (the internal microphone) were not
visible to root anymore, so the LEDs were always being set off.

Checking the output of alsamixer showed that alsa for my user account
was using the PulseAudio backend, but the root account doesn't have
access to my PulseAudio server and so falls back to directly accessing
the hardware instead. Checking freedesktop wiki pages suggested that
root should never really run its own PulseAudio server (and even
System-Wide mode is discouraged), so I'm not sure how root could see the
Capture controller in the past but I won't be enabling access to a
PulseAudio server to make it visible again.

Instead, this change moves the logic for checking the current status of
the microphone to my user account and only uses the sudo-run scripts to
actually set the LEDs.