Move udev rule file to avoid symlink
This change is necessary because udev starts early in the boot process
before symlinks to another partition can be resolved properly. The udev
rule in this repo is now only a backup file which is used during
editing, and the real rule is a normal file in the usual rules location
location instead of a symlink set up by stow.