Rework xkb shortcuts to avoid manual evdev copy
Previously I created a copy of the evdev keycode file and manually
chaged the keycodes, which I updated whenever the original evdev keycode
file was changed by upstream. Additionally, users are supposed to modify
keysyms, not keycodes. This change removes the custom xkb keycode file
in favor of an xkb types file and an xkb symbols file. These new files
have the additional advantage of extending the default configuration as
opposed to completely replacing it. Finally, the Xkbmap dotfile has been
updated to use the new xkb types and symbols. Though xkbcomp was
recommended by many online, staying with setxkbmap allows the setup to
continue working without modification of how lightdm starts the X
session (as seen in /etc/lightdm/Xsession). Since options passed to the
setxkbmap command completely replace those provided by the evdev rules,
all types and symbols (not just the custom ones in the new files) are
specified in the Xkbmap dotfile.