Simplify auth process for OSU student SSH server
OSU requires an initial password login to the SSH servers before pubkey
auth is utilized (until an unknown timeout, after which a password login
is required again). However, this initial password login is only
recorded per-host, and the previous address for the student SSH server
was to a load balancer. This meant that successive connections could
occur to different hosts, causing multiple password logins. This change
picks a specific host to use so that only 1 initial password login is
needed at the expensive of giving up load balancing.