Ensure host options can override globals
Previously, the globals were listed outside of a host specifier at the
beginning of the file. However, as the ssh_config(5) man page states,
the first obtained value for each parameter is used, so these global
options couldn't be overridden by specific hosts. To fix this, I have
placed the global options in a wildcard host section at the end of the
file. This method is the recommended configuration for global options,
plus leaving them outside of a host specifier at the end of the file
meant that the options didn't get applied (as confirmed with ssh -v).