zsh: Add alias for privileged binary redirection
The tee command is helpful to redirect output from a user command to a
location which requires other privileges. However, tee also duplicates
the output to STDOUT, which in the case of binary output can mess with
the terminal display. To avoid that issue, this alias redirects the
STDOUT of tee to /dev/null.